+++ server/routes/articles.ts
@@ -28,3 +28,25 @@ articlesRouter.post(
     res.status(201).json(result)
   }),
 )
+
+articlesRouter.get(
+  '/:id',
+  wrap(async (req, res) => {
+    const id: string = req.params.id || ''
+
+    if (!id) {
+      res.sendStatus(400)
+      return
+    }
+
+    const mgr = getManager()
+    const result = await mgr.findOne(ArticleEntity, { id })
+
+    if (!result) {
+      res.sendStatus(404)
+      return
+    }
+
+    res.status(200).json(result)
+  }),
+)
